requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
pymongo>=4.6.0
dnspython>=2.4.2
redis>=5.0.0\n